third_party/blink/web_tests/http/tests/navigation/forward-to-fragment-fires-onload-expected.txt

CONSOLE MESSAGE: Starting test.
CONSOLE MESSAGE: 1. Got to page 2.  Going to #foo.
CONSOLE MESSAGE: 2. Got to page 2 #foo.  Going back to start.
CONSOLE MESSAGE: 3. Got back to start.  Going forward to page 2.
CONSOLE MESSAGE: 4. Got back to page 2.  Immediately going forward.
CONSOLE MESSAGE: 5. Got to page 2 onload.  Success.
Tests that loading is not stopped by going forward to a fragment.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

This test checks the behavior of loading on back navigation. Back-forward cache needs to be disabled to test the behavior, because otherwise we won't trigger a new load on back navigation.

PASS successfullyParsed is true

TEST COMPLETE
